Start testing line intersection
Remembering that positive and negative infinity are not the same.
And that the intersection of two lines can be a line, a point, or nothing at all. And in this case, reify 'empty', because we want to be '7' compatible, and I've finally found an example where 'empty' is part of that family.